@@ -8,7 +8,7 @@ import { HeaderComponent } from './components/header/header.component';
 import { HomeComponent } from './components/pages/home/home.component';
 
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { SharedModule } from './shared/shared.module';
 import { ToastModule } from 'primeng/toast';
 import { MessageService } from 'primeng/api';
@@ -23,33 +23,33 @@ import { FirstletterPipe } from './pipes/firstletter.pipe';
 import { TokenInjectHttpInterceptor } from './interceptors/TokenInjectHttpInterceptor';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    HomeComponent,
-    SpinnerComponent,
-    FirstletterPipe
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    AppRoutingModule,
-    FormsModule,
-    ReactiveFormsModule,
-    HttpClientModule,
-    //PRIME NG
-    ToastModule,
-    AvatarModule,
-    BadgeModule,
-    PaginatorModule,
-    //Shared
-    SharedModule,
-  ],
-  providers: [
-    MessageService,
-  { provide: HTTP_INTERCEPTORS, useClass: LoadingInterceptor, multi: true },
-  { provide: HTTP_INTERCEPTORS, useClass: TokenInjectHttpInterceptor, multi: true }
-  ],
-  bootstrap: [AppComponent]
-})
+    declarations: [
+        AppComponent,
+        HeaderComponent,
+        HomeComponent,
+        SpinnerComponent,
+        FirstletterPipe
+    ],
+    imports: [
+        BrowserModule,
+        BrowserAnimationsModule,
+        AppRoutingModule,
+        FormsModule,
+        ReactiveFormsModule,
+        //PRIME NG
+        ToastModule,
+        AvatarModule,
+        BadgeModule,
+        PaginatorModule,
+        //Shared
+        SharedModule
+    ],
+    providers: [
+        MessageService,
+        { provide: HTTP_INTERCEPTORS, useClass: LoadingInterceptor, multi: true },
+        { provide: HTTP_INTERCEPTORS, useClass: TokenInjectHttpInterceptor, multi: true },
+        provideHttpClient(withInterceptorsFromDi())
+    ],
+    bootstrap: [AppComponent]
+ })
 export class AppModule { }
